Adds paybill functionality, rewrites transaction file outputting
Apparently we somehow collectively completely forgot about paybill until this point.
Also, I reworked a LOT of the transaction file outputting, not just to make the convention more elegant - which I'll admit is an optimization, but it's also good practice, which is more imperative - but also fixes a lot of issues with transaction file output.